added RxJS to LotkaVolterra, split out tests
