+++ index.js
@@ -1,8 +1,8 @@
-import Diversity from './src/lib/Diversity';
-import Sigmoid from './src/lib/Sigmoid';
-import LotkaVolterra from './src/lib/Lotka-Volterra';
-import NonSteadyState from './src/lib/NonSteadyState';
-import SteadyState from './src/lib/SteadyState';
+import Diversity from './src/Diversity';
+import Sigmoid from './src/Sigmoid';
+import LotkaVolterra from './src/Lotka-Volterra';
+import NonSteadyState from './src/NonSteadyState';
+import SteadyState from './src/SteadyState';
 
 export default {
   Diversity,
