+++ package.json
@@ -10,16 +10,16 @@
     "@babel/core": "^7.0.0-beta.46",
     "@babel/preset-env": "^7.0.0-beta.46",
     "babel-loader": "^8.0.0-beta.2",
-    "chai": "^4.1.0",
+    "chai": "^4.1.2",
     "eslint": "^3.19.0",
     "eslint-config-airbnb-base": "^11.2.0",
     "eslint-plugin-import": "^2.7.0",
-    "mocha": "^3.2.0",
+    "mocha": "^5.1.1",
     "webpack": "^4.6.0",
     "webpack-cli": "^2.1.2"
   },
   "scripts": {
-    "test": "node_modules/.bin/mocha",
+    "test": "mocha",
     "webpack": "webpack --watch --progress"
   },
   "repository": {
@@ -37,6 +37,6 @@
   },
   "homepage": "https://github.com/blmgeo/eco-model#readme",
   "dependencies": {
-    "functional.js": "^0.8.0"
+    "rxjs": "^6.1.0"
   }
 }
